refactor: move ParamGuardedSelfWrites to core (was extensions/rails)
The module is pure Prism AST analysis with no CurrentAttributes or Rails
knowledge — given any method def + a param name, it extracts the
`self.<attr> = <param>.<m>` writes guaranteed when the param is non-nil.
Its old name (`TransitiveConstantWrites`) and location (extensions/rails)
described its sole consumer, not its capability, falsely implying
coupling. Moved to lib/rbs_infer/param_guarded_self_writes.rb and renamed
RbsInfer::ParamGuardedSelfWrites; the CurrentAttributes example stays as
doc illustration. No behavior change; unit 451.
